main.py renamed to minecraft.py This is to clear up confusion between the entry point
